Update migrations to exclude non-django managed db tables

+++ peers/migrations/0001_initial.py
@@ -3,34 +3,38 @@ import datetime
 from south.db import db
 from south.v2 import SchemaMigration
 from django.db import models
+from django.conf import settings
 
 
 class Migration(SchemaMigration):
 
     def forwards(self, orm):
         # Adding model 'PeerRange'
-        db.create_table(u'peer_range', (
-            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
-            ('network', self.gf('django.db.models.fields.CharField')(max_length=128)),
-        ))
-        db.send_create_signal('peers', ['PeerRange'])
+        if settings.PEER_RANGE_MANAGED_TABLE:
+            db.create_table(u'peer_range', (
+                ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+                ('network', self.gf('django.db.models.fields.CharField')(max_length=128)),
+            ))
+            db.send_create_signal('peers', ['PeerRange'])
 
         # Adding model 'TechcEmail'
-        db.create_table('techc_email', (
-            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
-            ('email', self.gf('django.db.models.fields.CharField')(max_length=352, db_column='email')),
-        ))
-        db.send_create_signal('peers', ['TechcEmail'])
+        if settings.PEER_TECHC_MANAGED_TABLE:
+            db.create_table('techc_email', (
+                ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+                ('email', self.gf('django.db.models.fields.CharField')(max_length=352, db_column='email')),
+            ))
+            db.send_create_signal('peers', ['TechcEmail'])
 
         # Adding model 'Peer'
-        db.create_table(u'peer', (
-            ('peer_id', self.gf('django.db.models.fields.IntegerField')(primary_key=True)),
-            ('peer_name', self.gf('django.db.models.fields.CharField')(max_length=128)),
-            ('peer_as', self.gf('django.db.models.fields.IntegerField')()),
-            ('peer_tag', self.gf('django.db.models.fields.CharField')(max_length=64)),
-            ('domain_name', self.gf('django.db.models.fields.CharField')(max_length=128, null=True, blank=True)),
-        ))
-        db.send_create_signal('peers', ['Peer'])
+        if settings.PEER_MANAGED_TABLE:
+            db.create_table(u'peer', (
+                ('peer_id', self.gf('django.db.models.fields.IntegerField')(primary_key=True)),
+                ('peer_name', self.gf('django.db.models.fields.CharField')(max_length=128)),
+                ('peer_as', self.gf('django.db.models.fields.IntegerField')()),
+                ('peer_tag', self.gf('django.db.models.fields.CharField')(max_length=64)),
+                ('domain_name', self.gf('django.db.models.fields.CharField')(max_length=128, null=True, blank=True)),
+            ))
+            db.send_create_signal('peers', ['Peer'])
 
         # Adding M2M table for field networks on 'Peer'
         db.create_table(u'peer_networks', (
@@ -48,23 +52,24 @@ class Migration(SchemaMigration):
         ))
         db.create_unique(u'peer_techc_emails', ['peer_id', 'techcemail_id'])
 
-
     def backwards(self, orm):
         # Deleting model 'PeerRange'
-        db.delete_table(u'peer_range')
+        if settings.PEER_RANGE_MANAGED_TABLE:
+            db.delete_table(u'peer_range')
 
         # Deleting model 'TechcEmail'
-        db.delete_table('techc_email')
+        if settings.PEER_TECHC_MANAGED_TABLE:
+            db.delete_table('techc_email')
 
         # Deleting model 'Peer'
-        db.delete_table(u'peer')
+        if settings.PEER_MANAGED_TABLE:
+            db.delete_table(u'peer')
 
         # Removing M2M table for field networks on 'Peer'
         db.delete_table('peer_networks')
 
         # Removing M2M table for field techc_emails on 'Peer'
         db.delete_table('peer_techc_emails')
-
 
     models = {
         'peers.peer': {
